requirements.txt
asgiref==3.8.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
creditcard==1.0.2
cryptography==45.0.5
Django==5.1.2
django-admin-display==1.3.0
django-adminlte3==0.1.6
django-cors-headers==4.7.0
django-credit-cards==0.4.1
django-extensions==4.1
django-filter==25.1
django-grappelli==4.0.2
django-jazzmin==3.0.1
django-jet-reboot==1.3.10
django-oauth-toolkit==3.0.1
django-simpleui==2025.6.24
django-suit @ https://github.com/darklow/django-suit/tarball/v2#sha256=336dc47db1955fe7451634a6aacf203887bd98d243e29ceeff7b54213fcf4e5f
django-unfold==0.64.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
environ==1.0
graphviz==0.21
gunicorn==23.0.0
idna==3.10
jwcrypto==1.5.6
oauthlib==3.3.1
packaging==25.0
pillow==11.1.0
psycopg==3.2.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pydot==4.0.1
pydotplus==2.0.2
PyJWT==2.9.0
pyparsing==3.2.3
python-decouple==3.8
python-slugify==8.0.4
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.1
text-unidecode==1.3
typing_extensions==4.13.1
tzdata==2024.2
urllib3==2.5.0
wheel==0.45.1
